+++ src/main/java/br/edu/fatecgru/mercado_inteligente/service/UsuarioService.java
@@ -26,7 +26,7 @@ public class UsuarioService {
 	}
 
 	// Listar pelo ID do usuário
-	public Usuario getById(int id) {
+	public Usuario getById(Long id) {
 		return usuarioRepository.findById(id).orElse(null);
 	}
